Excel to Forms Quiz/requirements.txt
pandas==1.5.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.27.0
openpyxl==3.0.10
